+++ examples/fixedKernel.cc
@@ -10,7 +10,7 @@
 using namespace std;
 
 int main() {
-    typedef float pixelType;
+    typedef double pixelType;
     
     double sigmaX = 2.0;
     double sigmaY = 2.5;
